update :: user interface 변경사항 적용
