+++ internal/app/user/usecase/get_ranking.go
@@ -17,9 +17,9 @@ func (uc *userUseCase) GetRanking(ctx context.Context, input *input.RankQueryInp
 	)
 
 	if input.UserType == constant.TypeStudent {
-		users, err = uc.UserRepository.FindStudentsWithFilter(ctx, input.SortType, input.Name, input.Class, input.Grade)
+		users, err = uc.UserRepository.RankStudents(ctx, input.SortType, input.Name, input.Class, input.Grade)
 	} else {
-		users, err = uc.UserRepository.FindNonStudentWithFilter(ctx, input.SortType, input.Name)
+		users, err = uc.UserRepository.RankNonStudents(ctx, input.SortType, input.Name)
 	}
 
 	if err != nil {
